BC-9277 - move some mappers from repo to domain folder
