+++ apps/server/src/modules/account/domain/mapper/account-idm-to-do.mapper.db.spec.ts
@@ -1,5 +1,6 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { Account, IdmAccount } from '../../domain';
+import { Account } from '../account';
+import { IdmAccount } from '../idm-account';
 import { AccountIdmToDoMapper } from './account-idm-to-do.mapper.abstract';
 import { AccountIdmToDoMapperDb } from './account-idm-to-do.mapper.db';
