+++ apps/server/src/modules/account/domain/mapper/account-idm-to-do.mapper.db.ts
@@ -1,13 +1,13 @@
-import { Account } from '../../domain/account';
-import { IdmAccount } from '../../domain/idm-account';
+import { Account } from '../account';
+import { IdmAccount } from '../idm-account';
 import { AccountIdmToDoMapper } from './account-idm-to-do.mapper.abstract';
 
-export class AccountIdmToDoMapperIdm extends AccountIdmToDoMapper {
+export class AccountIdmToDoMapperDb extends AccountIdmToDoMapper {
 	public mapToDo(account: IdmAccount): Account {
 		const createdDate = account.createdDate ? account.createdDate : new Date();
 		return new Account({
-			id: account.id,
-			idmReferenceId: undefined,
+			id: account.attDbcAccountId ?? '',
+			idmReferenceId: account.id,
 			userId: account.attDbcUserId,
 			systemId: account.attDbcSystemId,
 			username: account.username ?? '',
