+++ apps/server/src/modules/account/domain/services/account-idm.service.integration.spec.ts
@@ -6,9 +6,9 @@ import { ObjectId } from '@mikro-orm/mongodb';
 import { ConfigModule } from '@nestjs/config';
 import { Test, TestingModule } from '@nestjs/testing';
 import { v1 } from 'uuid';
-import { AccountIdmToDoMapper, AccountIdmToDoMapperDb } from '../../repo';
 import { AccountSave } from '../account-save';
 import { IdmAccount } from '../idm-account';
+import { AccountIdmToDoMapper, AccountIdmToDoMapperDb } from '../mapper';
 import { AccountServiceIdm } from './account-idm.service';
 import { AbstractAccountService } from './account.service.abstract';
