+++ apps/server/src/modules/account/domain/services/account.service.integration.spec.ts
@@ -12,12 +12,13 @@ import { Test, TestingModule } from '@nestjs/testing';
 import { cleanupCollections } from '@testing/cleanup-collections';
 import { MongoMemoryDatabaseModule } from '@testing/database';
 import { v1 } from 'uuid';
-import { AccountEntity, AccountIdmToDoMapper, AccountIdmToDoMapperDb } from '../../repo';
+import { AccountEntity } from '../../repo';
 import { accountFactory } from '../../testing';
 import { Account } from '../account';
 import { AccountSave } from '../account-save';
 import { IdmAccount } from '../idm-account';
 import { ACCOUNT_REPO, AccountRepo } from '../interface';
+import { AccountIdmToDoMapper, AccountIdmToDoMapperDb } from '../mapper';
 import { AccountServiceDb } from './account-db.service';
 import { AccountServiceIdm } from './account-idm.service';
 import { AccountService } from './account.service';
