+++ apps/server/src/modules/account/repo/index.ts
@@ -2,6 +2,3 @@ export { AccountEntity, IdmAccountProperties } from './account.entity';
 export { AccountMikroOrmRepo } from './account.repo';
 export { AccountDoToEntityMapper } from './mapper/account-do-to-entity.mapper';
 export { AccountEntityToDoMapper } from './mapper/account-entity-to-do.mapper';
-export { AccountIdmToDoMapper } from './mapper/account-idm-to-do.mapper.abstract';
-export { AccountIdmToDoMapperDb } from './mapper/account-idm-to-do.mapper.db';
-export { AccountIdmToDoMapperIdm } from './mapper/account-idm-to-do.mapper.idm';
