+++ apps/server/src/modules/account/repo/mapper/index.ts
@@ -1,4 +1,2 @@
-export * from './account-idm-to-do.mapper.abstract';
-export * from './account-idm-to-do.mapper.idm';
-export * from './account-idm-to-do.mapper.db';
+export * from './account-do-to-entity.mapper';
 export * from './account-entity-to-do.mapper';
